+++ core_image_loader/build.gradle
@@ -2,11 +2,10 @@ plugins {
     id 'kotlin-android'
     id 'kotlin-kapt'
     id 'com.android.library'
-    id 'org.jetbrains.kotlin.android'
 }
 
 android {
-    namespace 'com.don.core_validation'
+    namespace 'com.don.core_image_loader'
     compileSdk 32
 
     defaultConfig {
@@ -44,4 +43,8 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.4'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
+
+    // glide
+    implementation 'com.github.bumptech.glide:glide:4.14.2'
+    kapt 'com.github.bumptech.glide:compiler:4.14.2'
 }
