+++ core_image_loader/src/main/res/drawable/ic_placeholder_square.xml
